requirements.txt
pandas
numpy
scikit-learn
Flask
gunicorn
flask-cors
python-dotenv
pymongo
requests
python-dateutil